move map_layer_base
