+++ src/actinia_core/rest/base/map_layer_base.py
